db restructure and updated progress log
