+++ server/.env
@@ -1,3 +1,4 @@
-DATABASE_NAME =diary_app
+DATABASE_NAME = diary_app
 DATABASE_USER= postgres
-DATABASE_PASSWORD =Codingstuff1!
+DATABASE_PASSWORD = Codingstuff1!
+PORT = 8080
