+++ server/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "server": "json-server -p3001 --watch db.json",
-    "dev": "nodemon ./db/index.js"
+    "dev": "nodemon ./index.js"
   },
   "author": "Group 10",
   "license": "ISC",
